Don't trust the port in the source address of requests. At least one client gets it wrong: always send to the client port for clients, and the server port for relays.

+++ src/dnsmasq.h
@@ -1053,8 +1053,8 @@ void make_duid(time_t now);
 
 /* rfc3315.c */
 #ifdef HAVE_DHCP6
-size_t dhcp6_reply(struct dhcp_context *context, int interface, char *iface_name,  
-		   struct in6_addr *fallback, size_t sz, int is_multicast, time_t now);
+unsigned short dhcp6_reply(struct dhcp_context *context, int interface, char *iface_name,  
+			   struct in6_addr *fallback, size_t sz, int is_multicast, time_t now);
 #endif
 
 /* dhcp-common.c */
